+++ tools/luban-provisioner/templates/source/luban-dagster-dbt-starrocks-code-location-source-template/{{cookiecutter.app_name}}/tests/test_definitions.py
@@ -1,15 +1,14 @@
 import os
-import pytest
-from dagster import load_assets_from_modules
+import importlib
 
-# Ensure Dagster doesn't try to actually connect to dbt or StarRocks during simple definition tests
+# Ensure Dagster doesn't try to connect to dbt or StarRocks during simple definition tests
 os.environ["DAGSTER_DBT_PARSE_PROJECT_ON_LOAD"] = "0"
 os.environ["STARROCKS_HOST"] = "mock_host"
 os.environ["STARROCKS_PORT"] = "9030"
 os.environ["STARROCKS_USER"] = "mock_user"
 os.environ["STARROCKS_PASSWORD"] = "mock_pass"
 
-from {{cookiecutter.package_name}} import defs
+defs = importlib.import_module("{{cookiecutter.package_name}}").defs
 
 
 def test_definitions_load():